+++ CHANGELOG.md
@@ -83,3 +83,7 @@
 
 ### 0.2.6
 - Fix missing default values in MD Documentation
+
+### 0.3.0
+- Added Logger handler stack, logger can be through constructor parameters ('settings.log'). Set false to disable logger.  
+- Added way to add Request and Response Handler functions from constructor paramters ('settings.requestHandler' && 'settings.responseHandler').